Add admin categories feature and update layouts and routes
This commit includes the following changes:
- New file: app/assets/stylesheets/admin/categories.scss
- New file: app/controllers/admin/categories_controller.rb
- New file: app/helpers/categories_helper.rb
- New file: app/views/admin/categories/index.html.erb
- New file: app/views/admin/categories/new.html.erb
- New file: app/views/admin/categories/show.html
- Modified file: app/views/layouts/_top_nav.html.erb
- Modified file: app/views/layouts/application.html.erb
- Modified file: config/routes.rb

The changes made add the admin categories feature to the application, including the necessary stylesheets, controller, helper, and view files. Additionally, the layouts and routes have been updated to accommodate the new feature. Please review the specific changes in each file for more details.